Add files via upload
Download All Files:

Download all the files from the GitHub repository to your computer.
Once downloaded, drag the entire folder to your Desktop for easier access if you even download an folder. This step is important to ensure everything works properly!
Read the README and License:

Before proceeding, make sure to read through the entire README to understand the usage and instructions.
Don’t forget to check the LICENSE file for important information about usage rights.
Run the MLHP - Github.bat File:

After everything is set up, locate the file MLHP - Github.bat and double-click to run it.
If the .bat file doesn't run because of permissions or security restrictions, follow the instructions below.
If the .bat File Doesn’t Work:

If the MLHP - Github.bat file doesn’t execute correctly due to being a batch file, you can manually create it:
Open Notepad on your computer.
Copy the entire code inside the .bat file and paste it into Notepad.
Save the file with the extension .bat (for example, MLHP - Github.bat).
Double click the new .bat file to run it.
Ignore MildLime.exe:

Do not worry about the MildLime.exe file it's not harmful, but it's part of the process. Simply ignore it.
Enjoy:

Once everything is set up and the .bat file is running, enjoy the experience!